src/backend/converter_lambda/requirements.txt
apple-music-python
boto3
python-dotenv
spotipy